+++ examples/react-redux-complex/app/index.js
@@ -23,8 +23,7 @@ let ReactDOM = require('react-dom')
 let Redux = require('redux')
 let Provider = require('react-redux').Provider
 let cherrytree = require('cherrytree')
-let { App } = require('./cherry-react')
-let { routerReducer, dispatchTransition } = require('./cherry-redux')
+let { Application, Router, routerReducer, dispatchTransition } = require('./neon-react-redux')
 let routes = require('./routes')
 
 // create a store
@@ -32,22 +31,61 @@ let reducers = Redux.combineReducers({ router: routerReducer })
 let store = Redux.createStore(reducers, window.devToolsExtension && window.devToolsExtension())
 
 // we'll be fetching some routes asynchronously
-let fetchAsyncRoutes = transition =>
-  Promise.all(transition.routes.map(route => route.options.async && route.options.async().then(component =>
-    route.options.component = component
-  )))
+let fetchAsyncRoutes = router => ({
+  next: transition => {
+    return Promise.all(transition.routes.map(route => {
+      let routeOptions = router.getRouteOptions(route.name)
+      return routeOptions.async && routeOptions.async().then(component =>
+        routeOptions.component = component
+      )
+    })
+  },
+  done: transition => {}
+})
+
+let cancelMessages = router => (transition, redirect, cancel) => {
+  if (transition.path === '/messages') {
+    console.log('CANCELLING')
+    cancel()
+  } else {
+    next(null, transition)
+  }
+}
+
+let redirectMessages = router => (transition, redirect, cancel) => {
+  if (transition.path === '/messages') {
+    redirect('/messages?abc=123')
+  }
+}
+
+let errorsToWarnings = router => ({
+  error: next => err => {
+    console.warn('ERRORED', err.stack)
+    throw err
+  }
+})
+
+let fail = router => transition => {
+  if (transition.path.indexOf('123') > -1) {
+    throw new Error('Access Denied!')
+  }
+}
 
 // create a router
-let router = cherrytree(routes, { log: true }, [
-  fetchAsyncRoutes,
-  dispatchTransition(store)
-])
+let middleware = [ errorsToWarnings, redirectMessages, fetchAsyncRoutes, fail, dispatchTransition(store) ]
+let router = cherrytree(routes, middleware, { log: true, pushState: false })
+
+router.start()
 
 // render the app
 let rootEl = document.querySelector('#app')
 ReactDOM.render(
-  <Provider store={store}>
-    <App router={router} />
-  </Provider>,
+  <Router router={router}>
+    <Provider store={store}>
+      <Application />
+    </Provider>
+  </Router>,
   rootEl
 )
+
+window.router = router