+++ examples/react-redux-complex/app/neon-react-redux/components/Router.js
@@ -1,0 +1,22 @@
+let React = require('react')
+let PropTypes = React.PropTypes
+
+module.exports = React.createClass({
+  propTypes: {
+    router: PropTypes.object.isRequired
+  },
+
+  childContextTypes: {
+    router: PropTypes.object
+  },
+
+  getChildContext () {
+    return {
+      router: this.props.router
+    }
+  },
+
+  render: function () {
+    return this.props.children
+  }
+})