+++ examples/react-redux-complex/app/neon-react-redux/index.js
@@ -1,0 +1,4 @@
+module.exports.Application = require('./components/Application')
+module.exports.Router = require('./components/Router')
+module.exports.routerReducer = require('./reducers/routerReducer')
+module.exports.dispatchTransition = require('./reducers/dispatchTransition')