+++ examples/react-redux-complex/app/neon-react-redux/reducers/dispatchTransition.js
@@ -1,0 +1,20 @@
+const dispatchTransition = store => router => ({
+  name: 'dispatchTransition',
+  next: next => transition => {
+    store.dispatch({
+      type: '@@neon-router/TRANSITION',
+      payload: {
+        routes: transition.routes.map(r => ({
+          name: r.name,
+          path: r.path,
+          params: r.params
+        })),
+        params: transition.params,
+        query: transition.query
+      }
+    })
+    next(null, transition)
+  }
+})
+
+module.exports = dispatchTransition