+++ examples/react-redux-complex/app/neon-react-redux/reducers/routerReducer.js
@@ -1,0 +1,11 @@
+const routerReducer = (state = {}, action) => {
+  switch (action.type) {
+    case '@@neon-router/TRANSITION':
+      let transition = action.payload
+      return transition
+    default:
+      return state
+  }
+}
+
+module.exports = routerReducer