+++ examples/react-redux-complex/app/routes.js
@@ -6,24 +6,36 @@ let Messages = asyncRoute(resolve => require(['./views/messages'], delay(resolve
 let Profile = require('./views/profile')
 let ProfileIndex = require('./views/profile_index')
 
+// module.exports = function routes (route) {
+//   // We can pass arbitrary options in the second argument of the route
+//   // function call. Because in this case we're using React, let's attach
+//   // the relevant components to each route.
+//   // Path is the only special option that is used to construct and
+//   // match URLs as well as extract URL parameters.
+//   route('application', {path: '/', component: Application, abstract: true}, function () {
+//     route('home', {path: '', component: Home})
+//     route('messages', {async: Messages})
+//     route('status', {path: ':user/status/:id'})
+//     route('profile', {path: ':user', component: Profile, abstract: true}, function () {
+//       route('profile.index', {component: ProfileIndex, path: ''})
+//       route('profile.lists')
+//       route('profile.edit')
+//     })
+//   })
+// }
 
-module.exports = function routes (route) {
-  // We can pass arbitrary options in the second argument of the route
-  // function call. Because in this case we're using React, let's attach
-  // the relevant components to each route.
-  // Path is the only special option that is used to construct and
-  // match URLs as well as extract URL parameters.
-  route('application', {path: '/', component: Application, abstract: true}, function () {
-    route('home', {path: '', component: Home})
-    route('messages', {async: Messages})
-    route('status', {path: ':user/status/:id'})
-    route('profile', {path: ':user', component: Profile, abstract: true}, function () {
-      route('profile.index', {component: ProfileIndex, path: ''})
-      route('profile.lists')
-      route('profile.edit')
-    })
-  })
-}
+module.exports = [
+  { name: 'application', path: '/', component: Application, abstract: true, children: [
+    { name: 'home', path: '', component: Home },
+    { name: 'messages', async: Messages },
+    { name: 'status', path: ':user/status/:id' },
+    { name: 'profile', path: ':user', component: Profile, abstract: true, children: [
+      { name: 'profile.index', path: '', component: ProfileIndex },
+      { name: 'profile.lists' },
+      { name: 'profile.edit' }
+    ]}
+  ]}
+]
 
 function delay (resolve) {
   return function (component) {