+++ examples/react-redux-complex/app/views/application.js
@@ -13,9 +13,9 @@ module.exports = React.createClass({
         <div className='App-header'>
           <h1>Application</h1>
           <ul className='Nav'>
-            <li className='Nav-item'><a href={this.context.router.generate('home')}>Home</a></li>
-            <li className='Nav-item'><a href={this.context.router.generate('messages')}>Messages</a></li>
-            <li className='Nav-item'><a href={this.context.router.generate('profile.index', {user: 'scrobblemuch'})}>Profile</a></li>
+            <li className='Nav-item'><a href={this.context.router.href('home')}>Home</a></li>
+            <li className='Nav-item'><a href={this.context.router.href('messages')}>Messages</a></li>
+            <li className='Nav-item'><a href={this.context.router.href('profile.index', {user: 'scrobblemuch'})}>Profile</a></li>
           </ul>
         </div>
         <div className='Container'>